Allow modal to set question title state
