+++ src/App.tsx
@@ -12,6 +12,7 @@ import {MenuModalRef} from "./screens/MenuModal";
 import {defaultPreview} from "./components/Preview";
 import {defaultCdn} from "./components/CDNUploadModal";
 import { RenameResult } from "./screens/RenameModal";
+import { Content } from "./isaac-data-types";
 
 
 export const AppContext = createContext({
@@ -24,7 +25,7 @@ export const AppContext = createContext({
     }) as NavigateFunction,
     menuModal: {current: null} as MutableRefObject<MenuModalRef | null>,
     showRenameModal: async (_: string) => null as RenameResult | null,
-    showPartTitleModal: (_?: boolean) => {},
+    showPartTitleModal: () => {},
     setActionRunning: (_: boolean) => {},
     preview: defaultPreview,
     cdn: defaultCdn,
