+++ src/screens/EditorScreen.tsx
@@ -32,6 +32,7 @@ import { QuestionTypes } from "../components/semantic/presenters/questionPresent
 import { formatTabIndex } from "../utils/keyedListHook";
 import { convertNumberToRoman } from "cr-numeral";
 import { closedPartTitleModalState, SetPartTitleModal, showPartTitleModal } from "./SetPartTitleModal";
+import { areQuestionTitlesMismatched } from "../utils/setQuestionTitles";
 
 const FILE_COMPONENTS = {
     "json": SemanticEditor,
@@ -139,69 +140,6 @@ export function EditorScreen() {
 
     const [actionRunning, setActionRunning] = useState(false);
 
-    const formatTitle = (accordionIndex: number, questionIndex: number, onlyQuestion?: boolean) => {
-        return `${formatTabIndex(accordionIndex, "alphabetical")}${onlyQuestion ? "" : `.${convertNumberToRoman(questionIndex + 1).toLowerCase()}`}`;
-    };
-
-    const isTitleStandardFormat = (title?: string) => {
-        if (!title) return true;
-
-        const regexFormat = new RegExp(/^[A-Z](?:\.(?=(?=[clxvi])(c{0,3}(?:xc|xl|l?x{0,3})(?:ix|iv|v?i{0,3})))\1)?$/);
-        return regexFormat.test(title);
-    };
-
-    const calculateQuestionTitles = useCallback((content: Content | string, writeNewTitles: boolean, overrideOldTitles: boolean) => {
-        if (typeof content === "string") return;
-
-        let noMismatch = true;
-        content.children?.forEach((child) => {
-            const contentChild = child as Content;
-            if (contentChild.type === "content" && contentChild.layout === "accordion") {
-                const accordion = child as Content;
-                accordion.children?.forEach((accordionChild, accordionIndex) => {
-                    const accordionSection = accordionChild as Content;
-                    const questionCount = accordionSection.children?.filter(c => (c.type || "") in QuestionTypes).length ?? 0;
-                    const inlineQuestionCount = accordionSection.children?.reduce((acc, c) => acc + (c.type === "isaacInlineRegion" ?
-                        (c as IsaacInlineQuestion).inlineQuestions?.length ?? 0 : 0), 0) ?? 0;
-                    const containsOneQuestion = questionCount + inlineQuestionCount === 1;
-
-                    let questionIndex = 0;
-                    accordionSection.children?.forEach((sectionChild) => {
-                        if ((sectionChild.type || "") in QuestionTypes) {
-                            const sectionQuestion = sectionChild as Content;
-                            console.log("I'm an accordion question: ", formatTitle(accordionIndex, questionIndex, containsOneQuestion), sectionQuestion.title, isTitleStandardFormat(sectionQuestion.title));
-                            if (writeNewTitles && (overrideOldTitles || isTitleStandardFormat(sectionQuestion.title))) {
-                                sectionQuestion.title = formatTitle(accordionIndex, questionIndex, containsOneQuestion);
-                            }
-                            noMismatch = noMismatch && sectionQuestion.title === formatTitle(accordionIndex, questionIndex, containsOneQuestion);
-                            questionIndex += 1;
-                        } else if (sectionChild.type === "isaacInlineRegion") {
-                            const inlineRegion = sectionChild as IsaacInlineQuestion;
-                            inlineRegion.inlineQuestions?.forEach((inlineQuestion) => {
-                                if ((inlineQuestion.type || "") in QuestionTypes) {
-                                    console.log("I'm an inline accordion question: ", formatTitle(accordionIndex, questionIndex, containsOneQuestion));
-                                    if (writeNewTitles && (overrideOldTitles || isTitleStandardFormat(inlineQuestion.title))) {
-                                        inlineQuestion.title = formatTitle(accordionIndex, questionIndex, containsOneQuestion);
-                                    }
-                                    noMismatch = noMismatch && inlineQuestion.title === formatTitle(accordionIndex, questionIndex, containsOneQuestion);
-                                    questionIndex += 1;
-                                } 
-                            });
-                        }
-                    });
-                });
-            }
-        });
-        if (writeNewTitles) {
-            setPartTitleMismatch(false);
-        } else {
-            setPartTitleMismatch(!noMismatch);
-        }
-
-        // Temporary for testing
-        setCurrentContent(content);
-    }, []);
-
     const setCurrentDoc = useCallback((content: Content | string, invertible = false) => {
         if (invertible) {
             const currentLastChanges = compare(currentContent, content, true);
@@ -209,8 +147,8 @@ export function EditorScreen() {
         }
         setCurrentContent(content);
         setDirty(hash(content) !== fileHash);
-        calculateQuestionTitles(content, true, false);
-    }, [fileHash, calculateQuestionTitles, currentContent]);
+        setPartTitleMismatch(areQuestionTitlesMismatched(content));
+    }, [fileHash, currentContent]);
     const loadNewDoc = useCallback((content: Content | string) => {
         setDirty(false);
         setFileHash(hash(content));
@@ -218,8 +156,8 @@ export function EditorScreen() {
         setIsAlreadyPublished(typeof content === "string" ? false : !!content.published);
         setCurrentContent(content);
         setCurrentContentPath(selection?.path);
-        calculateQuestionTitles(content, false, false);
-    }, [calculateQuestionTitles, selection?.path]);
+        setPartTitleMismatch(areQuestionTitlesMismatched(content));
+    }, [selection?.path]);
 
     const appContext = useMemo<ContextType<typeof AppContext>>(() => {
         async function dispatch(action: Action) {
